Initial commit of .emac.d.  Move me to poshrc repo
This is a WIP of setting up my own base configuration of Emacs. I have
been playing with Spacemacs but want to take a detour to experiment
with more minimalist setups again.

-----------

This is already in my poshrc repo. Today when attempting to get magit
to install through melpa I kept getting these types of errors:
   No such file or directory ...///magit-status.el[:Error:]
I lost the actual error but it was similar to the above.

Attempts:
1. Install with package.el through elpa -- probably failed for same
reason as all other attempts?  I verified that the files were in the
elpa directory.

** One interesting thing that I found through all these attempts were
that some parts of magit still worked if you ran it right after
installing before restarting Emacs. **

2. Install with package.el through melpa got [:Error:]
3. Reinstalled through package.el got [:Error:]
4. Reinstalled with package-reinstall got [:Error:]

- Thought about digging into the setup for magit in Emacs to see how
the file paths were getting messed up but decided to first check to
see if maybe there was something that did not like the symlinks in
Windows 11. I may go down that path or just make this repo into a
submodule of the poshrc instead.

5. Deleted symlink to .emacs.d and put a copy of the .emacs.d
directory in its place. Reinstalled magit and it works.
6. Tried readding the symlink and reinstalling magit. Same thing,
parts of magit work such as magit-status and magit-version but some
transient commands fail. Also after restarting Emacs magit then has
the [:Error:]